Changed all architecture settings (except for the sandbox and test targets) to ppc, i386, x86_64 in Development/Debug and $ARCHS_STANDARD_32_64_BIT in Deployment/Release. This mirrors the architecture settings of Growl.
Building ppc64 is pointless (especially for a debug build), since you'll never use it, so we omit it. Building x86_64 is pointless on Leopard (unless you're an application), but may not be on a future version of Mac OS X, so any hypothetical users of future versions of Mac OS X may want to have it built. That's why we (Growl) build x86_64 and not ppc64 in debug configurations. For release configurations, we assume it'll be sorted out in Xcode in a future release.