Allow mach_override to patch empty function bodies.
On 10.5, malloc_error_break is a whopping five instructions long:
0x01e314ae <malloc_error_break+0>:      push   %ebp
0x01e314af <malloc_error_break+1>:      mov    %esp,%ebp
0x01e314b1 <malloc_error_break+3>:      leave
0x01e314b2 <malloc_error_break+4>:      ret

If you want to patch this, the normal facilities by which mach_override works
will not hook this because it does not eat leave/ret, for good reason.

This adds the whole body of the above assembly to the beginning of
possibleInstructions so that if you override a method like this, it succeeds.

On 10.6 and 10.7, the assembly has changed to include two nop instructions,
making it large enough to override.